requirements.txt
Django>=6.0.2
requests>=2.32.5
beautifulsoup4>=4.14.3
psycopg2-binary>=2.9.11
python-dotenv>=1.2.2
hypothesis==6.151.9
django-q2>=1.9.0
rapidfuzz>=3.14.3
pillow>=12.1.1
